refact: reorganize HTTP method handling by separating methods into individual files for better maintainability
